Implement (optional) double validation
Add the option to split the validation set in half, and compute the
validation accuracy on the second half at the end of training.

This can be used to keep a "clean" validation split that is not
optimised over by the hyperparameter sweep, but can be used to decide
when to terminate it.